refactor: rename event files for consistency and update import paths
