+++ src/application/commands/common/Ping.ts
