+++ src/application/events/system/AppReady.ts
@@ -1,5 +1,5 @@
 import { Logger } from "@logging/logger";
-import { Event } from "@interfaces/discord/event";
+import { Event } from "@interfaces/discord/Event";
 import { container } from "tsyringe";
 
 export default new Event({
