+++ src/domain/interfaces/discord/Event.ts
